--- internal/setuporchestration/orchestrator_linux.go
@@ -1,975 +1,0 @@
-// SPDX-FileCopyrightText:  © 2025 Siemens Healthineers AG
-// SPDX-License-Identifier:   MIT
-
-//go:build linux
-
-package setuporchestration
-
-import (
-	"bytes"
-	"fmt"
-	"log/slog"
-	"net"
-	"os"
-	"os/exec"
-	"os/user"
-	"path/filepath"
-	"strings"
-	"time"
-
-	"github.com/siemens-healthineers/k2s/internal/core/config"
-)
-
-const (
-	flannelTemplateRelPath = "lib/modules/windows/node/k2s.node.module/linuxnode/distros/containernetwork/masternode/flannel.template.yml"
-	podNetworkCIDR         = "172.20.0.0/16"
-	servicesCIDR           = "172.21.0.0/16"
-	clusterDNS             = "172.21.0.10"
-	flannelNetworkName     = "cbr0"
-	flannelBackendType     = "vxlan"
-	kubeconfigSrc          = "/etc/kubernetes/admin.conf"
-	winVMName              = "k2s-win-worker"
-)
-
-const (
-	controlPlaneNodeNamesJSONPath = `jsonpath={range .items[*]}{.metadata.name}{"\n"}{end}`
-	nodeTaintKeysJSONPath         = `jsonpath={range .spec.taints[*]}{.key}{"\n"}{end}`
-)
-
-// LinuxOrchestrator implements Orchestrator using native Linux tools:
-// kubeadm, systemctl, and libvirt for managing a Windows worker VM.
-type LinuxOrchestrator struct{}
-
-// NewOrchestrator returns the platform-specific orchestrator.
-// On Linux, it returns a native kubeadm/systemd/libvirt orchestrator.
-func NewOrchestrator(_ interface{}) Orchestrator {
-	return &LinuxOrchestrator{}
-}
-
-func (o *LinuxOrchestrator) Install(cfg InstallConfig) error {
-	slog.Info("[Install] Installing K2s on Linux host", "linuxOnly", cfg.LinuxOnly)
-
-	if !cfg.LinuxOnly {
-		return fmt.Errorf("Linux host installation currently supports only --linux-only; Windows worker provisioning is not supported yet")
-	}
-	registryToken, err := readRegistryToken(cfg.InstallDir)
-	if err != nil {
-		return err
-	}
-
-	// Step 1: Validate the host before Kubernetes packages exist.
-	if err := o.checkHostPrerequisites(cfg); err != nil {
-		return fmt.Errorf("prerequisite check failed: %w", err)
-	}
-
-	// Step 2: Install the version-pinned Kubernetes and CRI-O package set.
-	k8sVersion, err := o.provisionKubernetes(cfg, registryToken)
-	if err != nil {
-		return fmt.Errorf("package provisioning failed: %w", err)
-	}
-
-	// Step 3: Install control plane natively via kubeadm.
-	if err := o.installControlPlane(cfg, k8sVersion); err != nil {
-		return fmt.Errorf("failed to install control plane: %w", err)
-	}
-
-	// Step 4: Set up kubeconfig for the user that invoked sudo.
-	if err := o.setupKubeconfig(); err != nil {
-		return fmt.Errorf("failed to setup kubeconfig: %w", err)
-	}
-
-	// Step 5: Deploy flannel CNI.
-	if err := o.deployFlannel(cfg); err != nil {
-		return fmt.Errorf("failed to deploy flannel: %w", err)
-	}
-	if err := o.deployClusterIPWebhook(cfg); err != nil {
-		return fmt.Errorf("failed to deploy ClusterIP allocation webhook: %w", err)
-	}
-	if err := o.installControlPlaneTools(cfg); err != nil {
-		return fmt.Errorf("failed to install native Linux control-plane tools: %w", err)
-	}
-
-	// Step 6: Wait for control plane node to be Ready.
-	if err := o.waitForNodeReady(120 * time.Second); err != nil {
-		slog.Warn("[Install] Control plane node not ready yet (may take a moment)", "error", err)
-	}
-	if err := o.removeControlPlaneTaints(); err != nil {
-		return fmt.Errorf("failed to make native Linux control-plane nodes schedulable: %w", err)
-	}
-	if err := o.configureHostDNS(cfg); err != nil {
-		return fmt.Errorf("failed to configure host Kubernetes DNS: %w", err)
-	}
-
-	// Step 7: Persist setup.json only after successful provisioning.
-	hostname, _ := os.Hostname()
-	clusterName := cfg.ClusterName
-	if clusterName == "" {
-		clusterName = "k2s-cluster"
-	}
-	if err := config.WriteRuntimeConfig(cfg.ConfigDir, "k2s", cfg.LinuxOnly, cfg.Version, clusterName, hostname, false); err != nil {
-		return fmt.Errorf("failed to write runtime config: %w", err)
-	}
-	if err := os.Chmod(filepath.Join(cfg.ConfigDir, "setup.json"), 0644); err != nil {
-		return fmt.Errorf("set runtime config permissions: %w", err)
-	}
-
-	if cfg.SkipStart {
-		if err := o.Stop(StopConfig{LinuxOnly: cfg.LinuxOnly, ConfigDir: cfg.ConfigDir}); err != nil {
-			return fmt.Errorf("stop cluster after --skip-start: %w", err)
-		}
-	}
-
-	slog.Info("[Install] K2s installation complete")
-	return nil
-}
-
-func (o *LinuxOrchestrator) Uninstall(cfg UninstallConfig) error {
-	slog.Info("[Uninstall] Uninstalling K2s from Linux host")
-
-	// Reset kubeadm
-	if !cfg.SkipPurge {
-		o.removeClusterIPWebhook()
-		_ = runCommand("systemctl", "stop", "kubelet")
-		if err := runCommand("kubeadm", "reset", "-f"); err != nil {
-			slog.Warn("[Uninstall] kubeadm reset failed (may already be clean)", "error", err)
-		}
-
-		o.cleanupCriOMirrorDropIns()
-		removeKubeletFileLogging()
-		o.removeControlPlaneState()
-	}
-
-	// Clean up network interfaces created by flannel
-	_ = runCommand("ip", "link", "delete", "cni0")
-	_ = runCommand("ip", "link", "delete", "flannel.1")
-
-	removeDNSProxy(cfg.ConfigDir)
-	removeHTTPProxy(cfg.LinuxOnly, cfg.ConfigDir)
-
-	// Remove kubeconfig
-	if u, err := invokingUser(); err == nil {
-		kubeconfigPath := filepath.Join(u.HomeDir, ".kube", "config")
-		if err := os.Remove(kubeconfigPath); err != nil && !os.IsNotExist(err) {
-			slog.Warn("[Uninstall] Could not remove kubeconfig", "path", kubeconfigPath, "error", err)
-		}
-	}
-
-	// Remove setup.json
-	if cfg.ConfigDir != "" {
-		if err := os.RemoveAll(cfg.ConfigDir); err != nil {
-			slog.Warn("[Uninstall] Could not remove config dir", "path", cfg.ConfigDir, "error", err)
-		}
-	}
-
-	slog.Info("[Uninstall] K2s uninstallation complete")
-	return nil
-}
-
-// removeControlPlaneState removes files created by kubeadm and Flannel. It is
-// deliberately limited to K2s control-plane paths so uninstall can recover an
-// interrupted Linux installation without changing unrelated host state.
-func (o *LinuxOrchestrator) removeControlPlaneState() {
-	for _, path := range []string{
-		"/etc/kubernetes",
-		"/var/lib/etcd",
-		"/var/lib/kubelet",
-		"/etc/cni/net.d/10-flannel.conflist",
-		"/run/flannel",
-	} {
-		if err := os.RemoveAll(path); err != nil && !os.IsNotExist(err) {
-			slog.Warn("[Uninstall] Could not remove control-plane state", "path", path, "error", err)
-		}
-	}
-}
-
-func (o *LinuxOrchestrator) cleanupCriOMirrorDropIns() {
-	removedAny := false
-
-	configPaths, err := filepath.Glob("/etc/containers/registries.conf.d/*.conf")
-	if err != nil {
-		slog.Warn("[Uninstall] Could not enumerate CRI-O registry configs", "error", err)
-		return
-	}
-
-	for _, configPath := range configPaths {
-		content, err := os.ReadFile(configPath)
-		if err != nil {
-			if !os.IsNotExist(err) {
-				slog.Warn("[Uninstall] Could not read CRI-O registry config", "path", configPath, "error", err)
-			}
-			continue
-		}
-
-		if !strings.Contains(string(content), "[[registry.mirror]]") {
-			continue
-		}
-
-		if err := os.Remove(configPath); err != nil {
-			if !os.IsNotExist(err) {
-				slog.Warn("[Uninstall] Could not remove CRI-O registry mirror config", "path", configPath, "error", err)
-			}
-			continue
-		}
-
-		removedAny = true
-		slog.Info("[Uninstall] Removed CRI-O registry mirror config", "path", configPath)
-	}
-
-	if !removedAny {
-		return
-	}
-
-	if err := runCommand("systemctl", "daemon-reload"); err != nil {
-		slog.Warn("[Uninstall] Could not reload systemd after CRI-O registry mirror cleanup", "error", err)
-	}
-
-	if err := runCommand("systemctl", "is-active", "--quiet", "crio"); err == nil {
-		if err := runCommand("systemctl", "restart", "crio"); err != nil {
-			slog.Warn("[Uninstall] Could not restart CRI-O after registry mirror cleanup", "error", err)
-		}
-	}
-}
-
-func (o *LinuxOrchestrator) Start(cfg StartConfig) error {
-	slog.Info("[Start] Starting K2s cluster on Linux host")
-
-	if cfg.LinuxOnly {
-		if err := runCommand("systemctl", "start", proxyNetworkSvc); err != nil {
-			return fmt.Errorf("failed to start Linux-only proxy compatibility network: %w", err)
-		}
-	}
-	if err := runCommand("systemctl", "start", proxyService); err != nil {
-		return fmt.Errorf("failed to start local HTTP proxy: %w", err)
-	}
-	if err := runCommand("systemctl", "start", crioServiceName); err != nil {
-		return fmt.Errorf("failed to start CRI-O: %w", err)
-	}
-
-	// Start kubelet — control plane static pods come up automatically.
-	if err := runCommand("systemctl", "start", "kubelet"); err != nil {
-		return fmt.Errorf("failed to start kubelet: %w", err)
-	}
-
-	// Wait for API server to be reachable
-	if err := o.waitForAPIServer(60 * time.Second); err != nil {
-		return fmt.Errorf("API server not reachable after start: %w", err)
-	}
-	if err := o.configureHostDNS(InstallConfig{InstallDir: cfg.InstallDir, ConfigDir: cfg.ConfigDir}); err != nil {
-		return fmt.Errorf("configure host Kubernetes DNS after start: %w", err)
-	}
-
-	slog.Info("[Start] K2s cluster started")
-	return nil
-}
-
-func (o *LinuxOrchestrator) Stop(cfg StopConfig) error {
-	slog.Info("[Stop] Stopping K2s cluster on Linux host")
-	if err := stopHostDNS(cfg.ConfigDir); err != nil {
-		return fmt.Errorf("restore host DNS before stopping K2s: %w", err)
-	}
-
-	// Stop kubelet
-	if err := runCommand("systemctl", "stop", "kubelet"); err != nil {
-		return fmt.Errorf("failed to stop kubelet: %w", err)
-	}
-	if err := runCommand("systemctl", "stop", crioServiceName); err != nil {
-		slog.Warn("[Stop] Could not stop CRI-O", "error", err)
-	}
-	if err := runCommand("systemctl", "stop", proxyService); err != nil {
-		slog.Warn("[Stop] Could not stop local HTTP proxy", "error", err)
-	}
-	if cfg.LinuxOnly {
-		if err := runCommand("systemctl", "stop", proxyNetworkSvc); err != nil {
-			slog.Warn("[Stop] Could not stop Linux-only proxy compatibility network", "error", err)
-		}
-	}
-
-	slog.Info("[Stop] K2s cluster stopped")
-	return nil
-}
-
-// ---------- control plane installation ----------
-
-func (o *LinuxOrchestrator) installControlPlane(cfg InstallConfig, k8sVersion string) error {
-	slog.Info("[Install] Installing Kubernetes control plane via kubeadm")
-
-	if err := runCommand("systemctl", "start", crioServiceName); err != nil {
-		return fmt.Errorf("failed to start CRI-O: %w", err)
-	}
-
-	// Generate kubeadm init config with KubeletConfiguration
-	initConfig := fmt.Sprintf(`apiVersion: kubeadm.k8s.io/v1beta4
-kind: InitConfiguration
-nodeRegistration:
-  criSocket: %s
----
-apiVersion: kubeadm.k8s.io/v1beta4
-kind: ClusterConfiguration
-kubernetesVersion: %s
-networking:
-  podSubnet: %s
-  serviceSubnet: %s
----
-apiVersion: kubelet.config.k8s.io/v1beta1
-kind: KubeletConfiguration
-failCgroupV1: false
-`, crioSocket, k8sVersion, podNetworkCIDR, servicesCIDR)
-
-	configDir := "/tmp/kubeadm-init"
-	configPath := filepath.Join(configDir, "kubeadm-init.yaml")
-
-	if err := os.MkdirAll(configDir, 0755); err != nil {
-		return fmt.Errorf("failed to create kubeadm init config directory: %w", err)
-	}
-	if err := os.WriteFile(configPath, []byte(initConfig), 0600); err != nil {
-		return fmt.Errorf("failed to write kubeadm init config: %w", err)
-	}
-	if cfg.ShowLogs {
-		slog.Info("[Install] Rendered kubeadm configuration", "path", configPath, "content", initConfig)
-	}
-
-	// Build kubeadm init arguments
-	args := []string{
-		"init",
-		"--config=" + configPath,
-	}
-
-	if err := runCommandWithLogs(cfg.ShowLogs, "kubeadm", args...); err != nil {
-		return fmt.Errorf("kubeadm init failed: %w", err)
-	}
-
-	// Clean up temporary config
-	os.Remove(configPath)
-
-	// Enable kubelet to start on boot
-	if err := runCommand("systemctl", "enable", "kubelet"); err != nil {
-		slog.Warn("[Install] Could not enable kubelet service", "error", err)
-	}
-
-	slog.Info("[Install] Control plane initialized")
-	return nil
-}
-
-// ---------- kubeconfig setup ----------
-
-func (o *LinuxOrchestrator) setupKubeconfig() error {
-	slog.Info("[Install] Setting up kubeconfig")
-
-	u, err := invokingUser()
-	if err != nil {
-		return fmt.Errorf("failed to determine current user: %w", err)
-	}
-
-	kubeDir := filepath.Join(u.HomeDir, ".kube")
-	if err := os.MkdirAll(kubeDir, 0755); err != nil {
-		return fmt.Errorf("failed to create .kube directory: %w", err)
-	}
-
-	destPath := filepath.Join(kubeDir, "config")
-
-	// Copy admin.conf to user's kubeconfig
-	input, err := os.ReadFile(kubeconfigSrc)
-	if err != nil {
-		return fmt.Errorf("failed to read kubeconfig from %s: %w", kubeconfigSrc, err)
-	}
-	if err := os.WriteFile(destPath, input, 0600); err != nil {
-		return fmt.Errorf("failed to write kubeconfig to %s: %w", destPath, err)
-	}
-	uid, err := lookupUserID(u.Uid)
-	if err != nil {
-		return err
-	}
-	gid, err := lookupUserID(u.Gid)
-	if err != nil {
-		return err
-	}
-	if err := os.Chown(kubeDir, uid, gid); err != nil {
-		return fmt.Errorf("set kubeconfig directory ownership: %w", err)
-	}
-	if err := os.Chown(destPath, uid, gid); err != nil {
-		return fmt.Errorf("set kubeconfig ownership: %w", err)
-	}
-
-	slog.Info("[Install] Kubeconfig written", "path", destPath)
-	return nil
-}
-
-func invokingUser() (*user.User, error) {
-	if sudoUser := os.Getenv("SUDO_USER"); sudoUser != "" && sudoUser != "root" {
-		return user.Lookup(sudoUser)
-	}
-	return user.Current()
-}
-
-// ---------- flannel deployment ----------
-
-func (o *LinuxOrchestrator) deployFlannel(cfg InstallConfig) error {
-	slog.Info("[Install] Deploying flannel CNI")
-
-	templatePath := filepath.Join(cfg.InstallDir, flannelTemplateRelPath)
-
-	templateBytes, err := os.ReadFile(templatePath)
-	if err != nil {
-		return fmt.Errorf("failed to read flannel template from %s: %w", templatePath, err)
-	}
-
-	// Replace template placeholders
-	manifest := string(templateBytes)
-	manifest = strings.ReplaceAll(manifest, "NETWORK.NAME", flannelNetworkName)
-	manifest = strings.ReplaceAll(manifest, "NETWORK.ADDRESS", podNetworkCIDR)
-	manifest = strings.ReplaceAll(manifest, "NETWORK.TYPE", flannelBackendType)
-
-	// Write rendered manifest to a temp file and apply
-	tmpFile, err := os.CreateTemp("", "k2s-flannel-*.yml")
-	if err != nil {
-		return fmt.Errorf("failed to create temp file for flannel manifest: %w", err)
-	}
-	defer os.Remove(tmpFile.Name())
-
-	if _, err := tmpFile.WriteString(manifest); err != nil {
-		tmpFile.Close()
-		return fmt.Errorf("failed to write flannel manifest: %w", err)
-	}
-	tmpFile.Close()
-
-	if err := runCommandWithLogs(cfg.ShowLogs, "kubectl", "--kubeconfig", kubeconfigSrc, "apply", "-f", tmpFile.Name()); err != nil {
-		return fmt.Errorf("failed to apply flannel manifest: %w", err)
-	}
-
-	slog.Info("[Install] Flannel deployed")
-	return nil
-}
-
-// installControlPlaneTools keeps native Linux feature parity with the
-// Windows-host kubemaster flow, which installs bundled CLI tools after
-// provisioning the Linux control plane.
-func (o *LinuxOrchestrator) installControlPlaneTools(cfg InstallConfig) error {
-	scriptPath := filepath.Join(cfg.InstallDir, "lib", "modules", "windows", "node", "k2s.node.module", "linuxnode", "distros", "scripts", "install-cli-tools.sh")
-	if _, err := os.Stat(scriptPath); err != nil {
-		return fmt.Errorf("required control-plane tools installer is missing at %s: %w", scriptPath, err)
-	}
-
-	slog.Info("[Install] Installing bundled CLI tools on native Linux control plane")
-	if err := runCommandWithLogs(cfg.ShowLogs, "bash", scriptPath); err != nil {
-		return fmt.Errorf("run CLI tools installer: %w", err)
-	}
-	for _, tool := range []string{"helm", "yq"} {
-		if _, err := exec.LookPath(tool); err != nil {
-			return fmt.Errorf("control-plane tool %q is unavailable after installation: %w", tool, err)
-		}
-	}
-	return nil
-}
-
-// deployClusterIPWebhook keeps native Linux aligned with the Windows-host
-// kubemaster topology: kubeadm allocates from the overall Service CIDR, while
-// the existing webhook assigns workload Services from OS-specific partitions.
-func (o *LinuxOrchestrator) deployClusterIPWebhook(cfg InstallConfig) error {
-	serviceCIDRs, err := config.ReadServiceCIDRConfig(cfg.InstallDir)
-	if err != nil {
-		return err
-	}
-	manifestDir := filepath.Join(cfg.InstallDir, "lib", "manifests", "clusterip-webhook")
-	for _, manifest := range []string{"namespace.yaml", "rbac.yaml", "webhook-config.yaml"} {
-		manifestPath := filepath.Join(manifestDir, manifest)
-		if err := runCommandWithLogs(cfg.ShowLogs, "kubectl", "--kubeconfig", kubeconfigSrc, "apply", "-f", manifestPath); err != nil {
-			return fmt.Errorf("apply ClusterIP webhook manifest %s: %w", manifest, err)
-		}
-	}
-
-	deploymentPath := filepath.Join(manifestDir, "deployment.yaml")
-	deployment, err := os.ReadFile(deploymentPath)
-	if err != nil {
-		return fmt.Errorf("read ClusterIP webhook deployment: %w", err)
-	}
-	renderedDeployment, err := renderClusterIPWebhookDeployment(string(deployment), serviceCIDRs)
-	if err != nil {
-		return err
-	}
-	renderedPath := filepath.Join("/tmp", "k2s-clusterip-webhook-deployment.yaml")
-	if err := os.WriteFile(renderedPath, []byte(renderedDeployment), 0600); err != nil {
-		return fmt.Errorf("write rendered ClusterIP webhook deployment: %w", err)
-	}
-	defer os.Remove(renderedPath)
-	if err := runCommandWithLogs(cfg.ShowLogs, "kubectl", "--kubeconfig", kubeconfigSrc, "apply", "-f", renderedPath); err != nil {
-		return fmt.Errorf("apply ClusterIP webhook deployment: %w", err)
-	}
-	if err := runCommand("kubectl", "--kubeconfig", kubeconfigSrc, "rollout", "status", "deployment/clusterip-webhook", "-n", "k2s-webhook", "--timeout=120s"); err != nil {
-		return fmt.Errorf("wait for ClusterIP webhook readiness: %w", err)
-	}
-	slog.Info("[Install] ClusterIP allocation webhook deployed", "linuxServiceCIDR", serviceCIDRs.LinuxCIDR, "windowsServiceCIDR", serviceCIDRs.WindowsCIDR)
-	return nil
-}
-
-func renderClusterIPWebhookDeployment(manifest string, serviceCIDRs config.ServiceCIDRConfig) (string, error) {
-	const defaultLinuxArgument = "--linux-subnet=172.21.0.0/24"
-	const defaultWindowsArgument = "--windows-subnet=172.21.1.0/24"
-	if strings.Count(manifest, defaultLinuxArgument) != 1 || strings.Count(manifest, defaultWindowsArgument) != 1 {
-		return "", fmt.Errorf("ClusterIP webhook deployment must contain exactly one Linux and one Windows subnet argument")
-	}
-
-	manifest = strings.Replace(manifest, defaultLinuxArgument, "--linux-subnet="+serviceCIDRs.LinuxCIDR, 1)
-	manifest = strings.Replace(manifest, defaultWindowsArgument, "--windows-subnet="+serviceCIDRs.WindowsCIDR, 1)
-	if !strings.Contains(manifest, "--linux-subnet="+serviceCIDRs.LinuxCIDR) || !strings.Contains(manifest, "--windows-subnet="+serviceCIDRs.WindowsCIDR) {
-		return "", fmt.Errorf("rendered ClusterIP webhook deployment does not contain configured service CIDRs")
-	}
-	return manifest, nil
-}
-
-func (o *LinuxOrchestrator) removeClusterIPWebhook() {
-	for _, resource := range []string{
-		"mutatingwebhookconfiguration/k2s-webhook",
-		"clusterrolebinding/k2s:clusterip-webhook",
-		"clusterrole/k2s:clusterip-webhook",
-		"namespace/k2s-webhook",
-	} {
-		if err := runCommand("kubectl", "--kubeconfig", kubeconfigSrc, "delete", resource, "--ignore-not-found=true", "--wait=false"); err != nil {
-			slog.Warn("[Uninstall] Could not remove ClusterIP webhook resource", "resource", resource, "error", err)
-		}
-	}
-}
-
-// ---------- readiness checks ----------
-
-func (o *LinuxOrchestrator) waitForNodeReady(timeout time.Duration) error {
-	slog.Info("[Install] Waiting for control plane node to be Ready", "timeout", timeout)
-
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		output, err := runCommandOutput("kubectl", "--kubeconfig", kubeconfigSrc, "get", "nodes", "-o", "jsonpath={.items[0].status.conditions[?(@.type=='Ready')].status}")
-		if err == nil && strings.TrimSpace(output) == "True" {
-			slog.Info("[Install] Control plane node is Ready")
-			return nil
-		}
-		time.Sleep(5 * time.Second)
-	}
-	return fmt.Errorf("control plane node not ready after %s", timeout)
-}
-
-func (o *LinuxOrchestrator) waitForAPIServer(timeout time.Duration) error {
-	slog.Info("[Start] Waiting for API server", "timeout", timeout)
-
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		if err := runCommand("kubectl", "--kubeconfig", kubeconfigSrc, "cluster-info"); err == nil {
-			return nil
-		}
-		time.Sleep(3 * time.Second)
-	}
-	return fmt.Errorf("API server not reachable after %s", timeout)
-}
-
-// removeControlPlaneTaints mirrors the Windows-host kubemaster setup: native
-// Linux control-plane nodes are also workload nodes. It removes only the
-// standard control-plane taint and verifies that it is absent afterwards.
-func (o *LinuxOrchestrator) removeControlPlaneTaints() error {
-	const controlPlaneTaint = "node-role.kubernetes.io/control-plane"
-
-	output, err := runCommandOutput("kubectl", "--kubeconfig", kubeconfigSrc, "get", "nodes", "-l", controlPlaneTaint, "-o", controlPlaneNodeNamesJSONPath)
-	if err != nil {
-		return fmt.Errorf("list control-plane nodes: %w", err)
-	}
-	nodeNames := nodeNamesFromKubectlOutput(output)
-	if len(nodeNames) == 0 {
-		return fmt.Errorf("no control-plane nodes found with label %q", controlPlaneTaint)
-	}
-
-	for _, nodeName := range nodeNames {
-		taints, err := runCommandOutput("kubectl", "--kubeconfig", kubeconfigSrc, "get", "node", nodeName, "-o", nodeTaintKeysJSONPath)
-		if err != nil {
-			return fmt.Errorf("verify taints on node %q: %w", nodeName, err)
-		}
-		if containsNodeTaint(taints, controlPlaneTaint) {
-			slog.Info("[Install] Removing control-plane taint to allow workloads", "node", nodeName)
-			if err := runCommand("kubectl", "--kubeconfig", kubeconfigSrc, "taint", "nodes", nodeName, controlPlaneTaint+"-"); err != nil {
-				return fmt.Errorf("remove control-plane taint from node %q: %w", nodeName, err)
-			}
-			taints, err = runCommandOutput("kubectl", "--kubeconfig", kubeconfigSrc, "get", "node", nodeName, "-o", nodeTaintKeysJSONPath)
-			if err != nil {
-				return fmt.Errorf("verify removed taints on node %q: %w", nodeName, err)
-			}
-		}
-		if containsNodeTaint(taints, controlPlaneTaint) {
-			return fmt.Errorf("control-plane taint remains on node %q", nodeName)
-		}
-	}
-
-	slog.Info("[Install] Native Linux control-plane nodes are schedulable", "count", len(nodeNames))
-	return nil
-}
-
-func nodeNamesFromKubectlOutput(output string) []string {
-	return strings.Fields(output)
-}
-
-func containsNodeTaint(output string, taint string) bool {
-	for _, value := range strings.Fields(output) {
-		if value == taint {
-			return true
-		}
-	}
-	return false
-}
-
-// ---------- Windows VM provisioning ----------
-
-func (o *LinuxOrchestrator) provisionWindowsVM(cfg InstallConfig) error {
-	slog.Info("[Install] Provisioning Windows worker VM via libvirt/KVM")
-
-	vmDataDir := filepath.Join(cfg.ConfigDir, "vm")
-	diskSizeGB := parseDiskSizeGB(cfg.MasterDiskSize, 50)
-
-	// Step 1: Create the K2s libvirt network
-	if err := CreateK2sNetwork(); err != nil {
-		return fmt.Errorf("failed to create K2s network: %w", err)
-	}
-
-	// Step 2: Prepare the Windows worker disk image (QCOW2)
-	diskPath, err := PrepareWindowsImage(cfg.InstallDir, vmDataDir, diskSizeGB)
-	if err != nil {
-		return fmt.Errorf("failed to prepare Windows worker image: %w", err)
-	}
-
-	// Step 3: Create and define the VM via libvirt
-	vmManager := NewVMManager()
-	cpuCount := parseCPUCount(cfg.MasterVMProcessorCount, 4)
-	memoryMB := parseMemoryMB(cfg.MasterVMMemory, 4096)
-
-	vmConfig := VMConfig{
-		Name:          winVMName,
-		ImagePath:     diskPath,
-		CPUCount:      cpuCount,
-		MemoryMB:      memoryMB,
-		DiskSizeGB:    diskSizeGB,
-		NetworkBridge: k2sNetworkName,
-	}
-
-	if err := vmManager.CreateVM(vmConfig); err != nil {
-		return fmt.Errorf("failed to create Windows VM: %w", err)
-	}
-
-	// Step 4: Start the VM
-	if err := vmManager.StartVM(winVMName); err != nil {
-		return fmt.Errorf("failed to start Windows VM: %w", err)
-	}
-
-	// Step 5: Wait for the VM to become reachable via SSH
-	slog.Info("[Install] Waiting for Windows VM to become reachable", "ip", winVMIP)
-	if err := waitForSSH(winVMIP, 22, 300*time.Second); err != nil {
-		return fmt.Errorf("Windows VM not reachable via SSH within timeout: %w", err)
-	}
-
-	// Step 6: Transfer K2s worker artifacts via SSH/SCP
-	slog.Info("[Install] Transferring K2s worker artifacts to Windows VM")
-	if err := transferWorkerArtifacts(cfg.InstallDir, winVMIP); err != nil {
-		return fmt.Errorf("failed to transfer worker artifacts: %w", err)
-	}
-
-	// Step 7: Install Windows services (containerd, kubelet, kube-proxy, flannel) via SSH
-	slog.Info("[Install] Installing K2s services on Windows VM")
-	if err := installWindowsServices(winVMIP); err != nil {
-		return fmt.Errorf("failed to install Windows services: %w", err)
-	}
-
-	// Step 8: Generate kubeadm join token and join the Windows node
-	slog.Info("[Install] Joining Windows node to cluster")
-	if err := joinWindowsNode(winVMIP); err != nil {
-		return fmt.Errorf("failed to join Windows node to cluster: %w", err)
-	}
-
-	// Step 9: Set up host routes for the Windows pod subnet
-	if err := SetupRoutes(); err != nil {
-		slog.Warn("[Install] Could not set up routes for Windows worker", "error", err)
-	}
-
-	// Step 10: Wait for Windows node to be Ready
-	slog.Info("[Install] Waiting for Windows node to be Ready")
-	if err := waitForWindowsNodeReady(180 * time.Second); err != nil {
-		slog.Warn("[Install] Windows node not ready yet (may take a moment)", "error", err)
-	}
-
-	slog.Info("[Install] Windows worker VM provisioned successfully")
-	return nil
-}
-
-// ---------- Windows VM helper functions ----------
-
-// waitForSSH polls the given host:port until an SSH connection can be established.
-func waitForSSH(host string, port int, timeout time.Duration) error {
-	deadline := time.Now().Add(timeout)
-	addr := fmt.Sprintf("%s:%d", host, port)
-
-	for time.Now().Before(deadline) {
-		conn, err := net.DialTimeout("tcp", addr, 5*time.Second)
-		if err == nil {
-			conn.Close()
-			slog.Info("[Install] SSH port reachable", "host", host)
-			return nil
-		}
-		time.Sleep(5 * time.Second)
-	}
-	return fmt.Errorf("SSH not reachable at %s after %s", addr, timeout)
-}
-
-// transferWorkerArtifacts copies K2s Windows worker binaries to the VM via SCP/SSH.
-func transferWorkerArtifacts(installDir, vmIP string) error {
-	// The artifacts to transfer are under installDir/bin/kube/ and installDir/bin/
-	// This includes: kubelet.exe, kubeadm.exe, kubectl.exe, kube-proxy.exe,
-	//                flannel.exe, containerd.exe, nssm.exe, nerdctl.exe, helm.exe
-	//
-	// For now, use a tar+SSH pipeline to bulk-transfer the bin/ directory.
-	// The Windows VM must have OpenSSH pre-installed in the base image.
-
-	binDir := filepath.Join(installDir, "bin")
-	remoteDir := `C:\k2s\bin`
-
-	// Create remote directory
-	if err := sshExecOnVM(vmIP, fmt.Sprintf(`mkdir -Force "%s"`, remoteDir)); err != nil {
-		slog.Warn("[Install] Could not create remote bin directory (may already exist)", "error", err)
-	}
-
-	// Transfer key binaries individually
-	binaries := []string{
-		"kube/kubelet.exe", "kube/kubeadm.exe", "kube/kubectl.exe", "kube/kube-proxy.exe",
-		"kube/flannel.exe", "kube/nssm.exe", "containerd/containerd.exe",
-		"nerdctl.exe", "kube/helm.exe",
-	}
-
-	for _, bin := range binaries {
-		localPath := filepath.Join(binDir, bin)
-		if _, err := os.Stat(localPath); err != nil {
-			slog.Warn("[Install] Binary not found, skipping", "path", localPath)
-			continue
-		}
-		remotePath := fmt.Sprintf(`C:\k2s\bin\%s`, filepath.Base(bin))
-		if err := scpToVM(vmIP, localPath, remotePath); err != nil {
-			return fmt.Errorf("failed to transfer '%s': %w", bin, err)
-		}
-	}
-
-	// Transfer configuration files
-	cfgDir := filepath.Join(installDir, "cfg")
-	if err := sshExecOnVM(vmIP, `mkdir -Force "C:\k2s\cfg"`); err != nil {
-		slog.Warn("[Install] Could not create remote cfg directory", "error", err)
-	}
-
-	// Create kubelet drop-in directory for configuration overrides
-	if err := sshExecOnVM(vmIP, `mkdir -Force "C:\etc\kubernetes\kubelet.conf.d"`); err != nil {
-		slog.Warn("[Install] Could not create kubelet drop-in directory", "error", err)
-	}
-
-	cfgFiles := []string{"kubeadm/joinnode.template.yaml", "containerd/config.toml", "cni/net-conf.json"}
-	for _, cf := range cfgFiles {
-		localPath := filepath.Join(cfgDir, cf)
-		if _, err := os.Stat(localPath); err != nil {
-			continue
-		}
-		remotePath := fmt.Sprintf(`C:\k2s\cfg\%s`, filepath.Base(cf))
-		if err := scpToVM(vmIP, localPath, remotePath); err != nil {
-			slog.Warn("[Install] Could not transfer config file", "file", cf, "error", err)
-		}
-	}
-
-	return nil
-}
-
-// installWindowsServices installs NSSM-managed services on the Windows VM.
-func installWindowsServices(vmIP string) error {
-	services := []struct {
-		name   string
-		binary string
-		args   string
-	}{
-		{"containerd", `C:\k2s\bin\containerd.exe`, `--config "C:\k2s\cfg\config.toml"`},
-		{"flanneld", `C:\k2s\bin\flannel.exe`, `--kubeconfig-file "C:\k2s\config" --iface=Ethernet --ip-masq --kube-subnet-mgr`},
-		{"kubelet", `C:\k2s\bin\kubelet.exe`, `--config "C:\k2s\cfg\kubelet-config.yaml" --config-dir "C:\etc\kubernetes\kubelet.conf.d" --kubeconfig "C:\k2s\config" --hostname-override=%COMPUTERNAME%`},
-		{"kubeproxy", `C:\k2s\bin\kube-proxy.exe`, `--kubeconfig "C:\k2s\config" --hostname-override=%COMPUTERNAME%`},
-	}
-
-	nssmPath := `C:\k2s\bin\nssm.exe`
-
-	for _, svc := range services {
-		cmd := fmt.Sprintf(`%s install %s "%s" %s`, nssmPath, svc.name, svc.binary, svc.args)
-		if err := sshExecOnVM(vmIP, cmd); err != nil {
-			return fmt.Errorf("failed to install service '%s': %w", svc.name, err)
-		}
-	}
-
-	// Start services in order
-	startOrder := []string{"containerd", "flanneld", "kubelet", "kubeproxy"}
-	for _, svc := range startOrder {
-		cmd := fmt.Sprintf(`%s start %s`, nssmPath, svc)
-		if err := sshExecOnVM(vmIP, cmd); err != nil {
-			slog.Warn("[Install] Could not start service (may be started by kubeadm join)", "service", svc, "error", err)
-		}
-	}
-
-	return nil
-}
-
-// joinWindowsNode generates a kubeadm join token and executes kubeadm join on the Windows VM.
-func joinWindowsNode(vmIP string) error {
-	// Generate a join command on the control plane
-	joinOutput, err := runCommandOutput("kubeadm", "token", "create", "--print-join-command")
-	if err != nil {
-		return fmt.Errorf("failed to create join token: %w", err)
-	}
-
-	joinCmd := strings.TrimSpace(joinOutput)
-	slog.Info("[Install] Join command generated", "command", joinCmd)
-
-	// Execute kubeadm join on the Windows VM
-	// Windows kubeadm expects: kubeadm join <api> --token <token> --discovery-token-ca-cert-hash <hash>
-	// --ignore-preflight-errors=IsPrivilegedUser,SystemVerification
-	winJoinCmd := fmt.Sprintf(`C:\k2s\bin\kubeadm.exe %s --ignore-preflight-errors=IsPrivilegedUser,SystemVerification --cri-socket npipe:////./pipe/containerd-containerd`,
-		strings.TrimPrefix(joinCmd, "kubeadm "))
-
-	if err := sshExecOnVM(vmIP, winJoinCmd); err != nil {
-		return fmt.Errorf("kubeadm join failed on Windows VM: %w", err)
-	}
-
-	return nil
-}
-
-// waitForWindowsNodeReady waits until the Windows worker node reports Ready.
-func waitForWindowsNodeReady(timeout time.Duration) error {
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		output, err := runCommandOutput("kubectl", "get", "nodes", "-o", "jsonpath={range .items[*]}{.metadata.name}={.status.conditions[?(@.type=='Ready')].status}{','}{end}")
-		if err == nil {
-			for _, entry := range strings.Split(output, ",") {
-				parts := strings.SplitN(entry, "=", 2)
-				if len(parts) == 2 && parts[1] == "True" && parts[0] != "" {
-					// Check if this is the Windows node (not the control plane)
-					nodeOS, _ := runCommandOutput("kubectl", "get", "node", parts[0], "-o", "jsonpath={.status.nodeInfo.operatingSystem}")
-					if strings.TrimSpace(nodeOS) == "windows" {
-						slog.Info("[Install] Windows node is Ready", "name", parts[0])
-						return nil
-					}
-				}
-			}
-		}
-		time.Sleep(5 * time.Second)
-	}
-	return fmt.Errorf("Windows node not ready after %s", timeout)
-}
-
-// sshExecOnVM executes a command on the Windows VM via SSH.
-// Requires OpenSSH to be installed on the Windows worker image.
-func sshExecOnVM(vmIP, command string) error {
-	slog.Debug("[SSH] Executing on Windows VM", "ip", vmIP, "command", command)
-	return runCommand("ssh",
-		"-o", "StrictHostKeyChecking=no",
-		"-o", "UserKnownHostsFile=/dev/null",
-		"-o", "ConnectTimeout=10",
-		fmt.Sprintf("remote@%s", vmIP),
-		command,
-	)
-}
-
-// scpToVM copies a local file to the Windows VM via SCP.
-func scpToVM(vmIP, localPath, remotePath string) error {
-	slog.Debug("[SCP] Copying to Windows VM", "local", localPath, "remote", remotePath, "ip", vmIP)
-	return runCommand("scp",
-		"-o", "StrictHostKeyChecking=no",
-		"-o", "UserKnownHostsFile=/dev/null",
-		localPath,
-		fmt.Sprintf("remote@%s:%s", vmIP, remotePath),
-	)
-}
-
-// ---------- parameter parsing helpers ----------
-
-func parseCPUCount(value string, defaultVal int) int {
-	if value == "" {
-		return defaultVal
-	}
-	n := defaultVal
-	fmt.Sscanf(value, "%d", &n)
-	if n <= 0 {
-		return defaultVal
-	}
-	return n
-}
-
-func parseMemoryMB(value string, defaultVal int) int {
-	if value == "" {
-		return defaultVal
-	}
-	val := strings.ToUpper(strings.TrimSpace(value))
-	n := defaultVal
-	if strings.HasSuffix(val, "GB") {
-		fmt.Sscanf(strings.TrimSuffix(val, "GB"), "%d", &n)
-		n *= 1024
-	} else if strings.HasSuffix(val, "MB") {
-		fmt.Sscanf(strings.TrimSuffix(val, "MB"), "%d", &n)
-	} else {
-		fmt.Sscanf(val, "%d", &n)
-	}
-	if n <= 0 {
-		return defaultVal
-	}
-	return n
-}
-
-func parseDiskSizeGB(value string, defaultVal int) int {
-	if value == "" {
-		return defaultVal
-	}
-	val := strings.ToUpper(strings.TrimSpace(value))
-	n := defaultVal
-	if strings.HasSuffix(val, "GB") {
-		fmt.Sscanf(strings.TrimSuffix(val, "GB"), "%d", &n)
-	} else {
-		fmt.Sscanf(val, "%d", &n)
-	}
-	if n <= 0 {
-		return defaultVal
-	}
-	return n
-}
-
-// ---------- command helpers ----------
-
-// runCommand executes a command and returns any error.
-func runCommand(name string, args ...string) error {
-	slog.Debug("Executing command", "cmd", name, "args", args)
-	cmd := exec.Command(name, args...)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		slog.Error("Command failed", "cmd", name, "output", string(output), "error", err)
-		return fmt.Errorf("%s failed: %w\nOutput: %s", name, err, string(output))
-	}
-	slog.Debug("Command succeeded", "cmd", name, "output", string(output))
-	return nil
-}
-
-// runCommandWithLogs streams command output when --output is requested while
-// retaining the command error context used by the normal execution path.
-func runCommandWithLogs(showLogs bool, name string, args ...string) error {
-	if !showLogs {
-		return runCommand(name, args...)
-	}
-
-	slog.Info("[Install] Running command", "cmd", name, "argumentCount", len(args))
-	cmd := exec.Command(name, args...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("%s failed: %w", name, err)
-	}
-	return nil
-}
-
-// runCommandOutput executes a command and returns only stdout. Stderr is kept
-// in failures so kubectl diagnostics do not corrupt parsed successful output.
-func runCommandOutput(name string, args ...string) (string, error) {
-	cmd := exec.Command(name, args...)
-	var stdout, stderr bytes.Buffer
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	if err := cmd.Run(); err != nil {
-		return "", fmt.Errorf("%s failed: %w; stderr: %s", name, err, strings.TrimSpace(stderr.String()))
-	}
-	return stdout.String(), nil
-}
